+++ R/extract_rules.R
@@ -9,25 +9,21 @@
 #' @param X The features matrix.
 #' @param ntrees The number of decision trees.
 #' @param ite_std The standardized ITE.
-#' @param take_1 Whether or not to call the take1 helper function.
 #' @param type_decay The type of decay to apply when pruning the rules.
+#' (1: relative error; 2: error)
 #'
 #' @keywords internal
 #'
 #' @return
 #' A vector of causal rules.
 #'
-extract_rules <- function(treelist, X, ntrees, ite_std, take_1, type_decay) {
+extract_rules <- function(treelist, X, ntrees, ite_std, type_decay) {
 
   rules <- inTrees::extractRules(treeList = treelist,
                                 X = X,
                                 ntree = ntrees,
                                 maxdepth = 15)
   rules <- c(rules)
-
-  if (take_1) {
-    rules <- rules[take1(length(rules))]
-  }
 
   rules_matrix <- matrix(rules)
   colnames(rules_matrix) <- "condition"